requirements.txt
Flask
flask-sqlalchemy
PyMySQL
python-dotenv
marshmallow
